Fixed API integration
Both API's now interact with other properly. Data is being moved and cleaned
