+++ extract.py
@@ -1,15 +1,2 @@
-from flask import Flask
-from prometheus_flask_exporter import PrometheusMetrics
-import PipelineSetup as pData
+import pandas as pd
 
-app = Flask(__name__)
-metrics = PrometheusMetrics(app, path='/metrics')
-
-@app.route('/')
-def index():
-    data = pData.pokeData().df
-    result = data.to_json(orient='values')
-    return result
-
-if __name__ == '__main__':
-    app.run()
